+++ pkg/api/controller.go
@@ -54,6 +54,9 @@ const (
 	// DefaultPerPage is the default number of results returned for paginated queries to the API
 	DefaultPerPage int = 100
 
+	defaultSTSTTLSeconds = 3600
+	maxSTSTTLSeconds     = 3600 * 12
+
 	lakeFSPrefix = "symlinks"
 
 	actionStatusCompleted = "completed"
@@ -536,7 +539,7 @@ func (c *Controller) Login(w http.ResponseWriter, r *http.Request, body apigen.L
 	expires := loginTime.Add(duration)
 	secret := c.Auth.SecretStore().SharedSecret()
 
-	tokenString, err := GenerateJWTLogin(secret, user.Username, loginTime.Unix(), expires.Unix())
+	tokenString, err := GenerateJWTLogin(secret, user.Username, loginTime, expires)
 	if err != nil {
 		writeError(w, r, http.StatusInternalServerError, http.StatusText(http.StatusInternalServerError))
 		return
@@ -556,22 +559,32 @@ func (c *Controller) Login(w http.ResponseWriter, r *http.Request, body apigen.L
 	writeResponse(w, r, http.StatusOK, response)
 }
 
-func (c *Controller) STSLogin(w http.ResponseWriter, r *http.Request, body apigen.STSLoginJSONRequestBody) {
+func (c *Controller) StsLogin(w http.ResponseWriter, r *http.Request, body apigen.StsLoginJSONRequestBody) {
 	ctx := r.Context()
-	responseData, err := c.Authentication.ValidateSTS(ctx, body.Code, body.RedirectUri, body.State)
+	externalUserID, err := c.Authentication.ValidateSTS(ctx, body.Code, body.RedirectUri, body.State)
 	if c.handleAPIError(ctx, w, r, err) {
 		return
 	}
 	// validate a user exists with the external user id
-	_, err = c.Auth.GetUserByExternalID(ctx, responseData.ExternalUserID)
+	_, err = c.Auth.GetUserByExternalID(ctx, externalUserID)
 	if c.handleAPIError(ctx, w, r, err) {
 		return
 	}
-	token, err := GenerateJWTLogin(c.Auth.SecretStore().SharedSecret(), responseData.ExternalUserID, time.Now().Unix(), responseData.ExpiresAtUnixTime)
+	expiresInSec := defaultSTSTTLSeconds
+	if body.TtlSeconds != nil {
+		expiresInSec = min(maxSTSTTLSeconds, int(*body.TtlSeconds))
+	}
+	now := time.Now()
+	expiresAt := now.Add(time.Duration(expiresInSec) * time.Second)
+	token, err := GenerateJWTLogin(c.Auth.SecretStore().SharedSecret(), externalUserID, now, expiresAt)
 	if c.handleAPIError(ctx, w, r, err) {
 		return
 	}
-	writeResponse(w, r, http.StatusOK, token)
+	responseToken := apigen.AuthenticationToken{
+		Token:           token,
+		TokenExpiration: swag.Int64(expiresAt.Unix()),
+	}
+	writeResponse(w, r, http.StatusOK, responseToken)
 }
 
 func (c *Controller) GetPhysicalAddress(w http.ResponseWriter, r *http.Request, repository, branch string, params apigen.GetPhysicalAddressParams) {
@@ -2638,7 +2651,8 @@ func (c *Controller) handleAPIErrorCallback(ctx context.Context, w http.Response
 		cb(w, r, http.StatusPreconditionFailed, "Precondition failed")
 	case errors.Is(err, authentication.ErrNotImplemented):
 		cb(w, r, http.StatusNotImplemented, "Not implemented")
-	case errors.Is(err, authentication.ErrInvalidSTS):
+	case errors.Is(err, authentication.ErrInsufficientPermissions):
+		c.Logger.WithContext(ctx).WithError(err).Info("User verification failed - insufficient permissions")
 		cb(w, r, http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized))
 	case err != nil:
 		c.Logger.WithContext(ctx).WithError(err).Error("API call returned status internal server error")